+++ docker.sh
@@ -94,4 +94,3 @@ case $cmd in
         echo "Unsupported command \"$cmd\""
         exit 1
 esac
-
