--- hardware/__init__.py
@@ -1,4 +1,0 @@
-from .robot_brains import RobotBrains
-from .testbrain_hardware import TestBrain
-
-__all__ = ['RobotBrains', 'TestBrain']
