+++ requirements-dev.txt
@@ -1,0 +1,5 @@
+-r requirements.txt
+autopep8
+mypy
+pre-commit
+pylint
