+++ test_brain/__init__.py
@@ -1,0 +1,5 @@
+from .system import System
+
+__all__ = [
+    'System',
+]
