+++ test_brain/hardware/__init__.py
@@ -1,0 +1,4 @@
+from .brain_configs import BrainConfigs
+from .test_brain import TestBrain
+
+__all__ = ['BrainConfigs', 'TestBrain']
